# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Flow API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/timesheet": {
            "get": {
                "tags": [
                    "Timesheet"
                ],
                "summary": "Get all timesheets with filtering",
                "operationId": "f054349ff2ccbb583632dd220b997eb0",
                "parameters": [
                    {
                        "name": "filters[user_id]",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "filters[date][gt]",
                        "in": "query",
                        "description": "Filter timesheets with date greater than",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-01"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered list of timesheets"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Timesheet"
                ],
                "summary": "Create a new timesheet",
                "operationId": "be1362d4d94156035295bcd25c8b4fbb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "project_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "task_name": {
                                        "type": "string",
                                        "example": "Worked on project backend"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-16"
                                    },
                                    "hours": {
                                        "type": "integer",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Timesheet added successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/timesheet/{id}": {
            "get": {
                "tags": [
                    "Timesheet"
                ],
                "summary": "Get timesheet by ID",
                "operationId": "f359eef0437cc86a909685e4b164ceb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Timesheet detail"
                    },
                    "404": {
                        "description": "Timesheet not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Timesheet"
                ],
                "summary": "Update timesheet details",
                "operationId": "cdd49a6a88d6b0105ab6e34f73dd4802",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "project_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "task_name": {
                                        "type": "string",
                                        "example": "Worked on project frontend"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-16"
                                    },
                                    "hours": {
                                        "type": "integer",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Timesheet updated successfully"
                    },
                    "404": {
                        "description": "Timesheet not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Timesheet"
                ],
                "summary": "Delete a timesheet",
                "operationId": "169a6be6dc04ddb38521c4f4ac98a5f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Timesheet deleted successfully"
                    },
                    "404": {
                        "description": "Timesheet not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: 'Bearer {token}'",
                "bearerFormat": "OAuth2",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "oauth2_security_example": [
                "read",
                "write"
            ],
            "passport": []
        }
    ]
}